// backend/src/types.ts
export type Program = {
  // TODO: We can make identifiers unique across all entity types and refactor to clean up grammar (wont have to include Entity other than in "create" statements ex. right now Entity is in update/delete/conditionals/animation/etc.)
  definitions: Persistence[]
  main: Execution
}

// -- Definitions --

export type Identifier = string
// ...
export enum EntityTypes {
  object = 'object',
  layer = 'layer',
  animation = 'animation',
}

export type Execution = Step[]

export type Step = Persistence | Delete | Animate | Global | Conditional

// --

export type Persistence = Create | Update

export type Parameter = { parameter: string; value: any }

export type Parameters =
  | Partial<eObject['parameters']>
  | Partial<eAnimation['parameters']>
  | Partial<eLayer['parameters']>

export type Create = {
  type: 'create'
  identifier: string
  entity: EntityTypes
  parameters: Parameters
}

export type Update = {
  type: 'update'
  identifier: string
  entity: EntityTypes
  parameters: Parameters
}

export type Delete = {
  type: 'delete'
  identifier: string
  entity: EntityTypes
}

// --

export type Animate = {
  type: 'animate'
  animationIdentifier: Identifier
  entity: EntityTypes.layer | EntityTypes.object
  entityIdentifier: Identifier
  action: Action
}

export enum Action {
  start = 'start',
  stop = 'stop',
  speedUp = 'speedUp',
  slowDown = 'slowDown',
  reset = 'reset',
}

// --

type Global = Collision | Wait

export type Collision = {
  type: 'collision'
  value: boolean
}

export type Wait = {
  type: 'wait'
  duration: number
}

// --

export type Conditional = {
  type: 'if' | 'loop'
  condition: Condition
  count?: number
  body: Step[]
}

export type OP = '<' | '>' | '<=' | '>=' | '==' | '!='

type Condition = {
  op: OP
  lhs: string
  rhs: string
}